digitalwallet_restapi/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
python-dotenv==1.0.1
websocket_client==1.6.1
psycopg2==2.9.6
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.3
asgiref==3.7.2